+++ test/sol/core/TestHeart1.sol
@@ -50,7 +50,7 @@ contract TestHeart1 {
         uint32 newSeconds = 15;
         heart.setReserveFreeze(assetCode, newSeconds);
 
-        uint32 reserveFreezeSecond = heart.getReserveFreeze(assetCode);
+        uint256 reserveFreezeSecond = heart.getReserveFreeze(assetCode);
 
         Assert.equal(uint(reserveFreezeSecond), uint(newSeconds), "heart.setReserveFreeze() should set reserve freeze seconds properly");
     }
@@ -62,7 +62,7 @@ contract TestHeart1 {
         Assert.equal(heart.getReserveFreeze(assetCode), uint(0), "heart.getReserveFreeze() should return reserve freeze second");
 
         heart.setReserveFreeze(assetCode, newSeconds);
-        uint32 reserveFreezeSecond = heart.getReserveFreeze(assetCode);
+        uint256 reserveFreezeSecond = heart.getReserveFreeze(assetCode);
 
         Assert.equal(uint(reserveFreezeSecond), uint(newSeconds), "heart.getReserveFreeze() should return reserve freeze second");
     }
